+++ src/app/cli/src/commands/verify_command.rs
@@ -9,7 +9,7 @@
 
 use std::sync::Arc;
 
-use futures::TryStreamExt;
+use futures::{StreamExt, TryStreamExt};
 use kamu::domain::*;
 use kamu::utils::datasets_filtering::filter_datasets_by_pattern;
 use opendatafabric::*;
@@ -25,16 +25,25 @@ type GenericVerificationResult = Result<Vec<VerificationResult>, CLIError>;
 pub struct VerifyCommand {
     dataset_repo: Arc<dyn DatasetRepository>,
     verification_svc: Arc<dyn VerificationService>,
+    dependency_graph_service: Arc<dyn DependencyGraphService>,
+    remote_alias_reg: Arc<dyn RemoteAliasesRegistry>,
     output_config: Arc<OutputConfig>,
     refs: Vec<DatasetRefPattern>,
     recursive: bool,
     integrity: bool,
 }
 
+struct RemoteRefDependency {
+    source_dataset: String,
+    dependencies: Vec<String>,
+}
+
 impl VerifyCommand {
     pub fn new<I>(
         dataset_repo: Arc<dyn DatasetRepository>,
         verification_svc: Arc<dyn VerificationService>,
+        dependency_graph_service: Arc<dyn DependencyGraphService>,
+        remote_alias_reg: Arc<dyn RemoteAliasesRegistry>,
         output_config: Arc<OutputConfig>,
         refs: I,
         recursive: bool,
@@ -46,6 +55,8 @@ impl VerifyCommand {
         Self {
             dataset_repo,
             verification_svc,
+            dependency_graph_service,
+            remote_alias_reg,
             output_config,
             refs: refs.collect(),
             recursive,
@@ -79,18 +90,62 @@ impl VerifyCommand {
     ) -> GenericVerificationResult {
         let dataset_ref_pattern = self.refs.first().unwrap();
 
-        let requests: Vec<_> = filter_datasets_by_pattern(
+        let dataset_ids: Vec<_> = filter_datasets_by_pattern(
             self.dataset_repo.as_ref(),
             vec![dataset_ref_pattern.clone()],
         )
-        .map_ok(|dataset_handle| VerificationRequest {
-            dataset_ref: dataset_handle.as_local_ref(),
-            block_range: (None, None),
-        })
+        .map_ok(|dataset_handle| dataset_handle.id)
         .try_collect()
         .await?;
 
-        if requests.is_empty() {
+        let requests: Vec<_> = if self.recursive {
+            self.dependency_graph_service
+                .get_recursive_upstream_dependencies(dataset_ids)
+                .await
+                .int_err()?
+                .map(|dataset_id| VerificationRequest {
+                    dataset_ref: DatasetRef::ID(dataset_id),
+                    block_range: (None, None),
+                })
+                .collect()
+                .await
+        } else {
+            dataset_ids
+                .iter()
+                .map(|dataset_id| VerificationRequest {
+                    dataset_ref: DatasetRef::ID(dataset_id.clone()),
+                    block_range: (None, None),
+                })
+                .collect()
+        };
+
+        let (filtered_requests, missed_remote_dependencies) =
+            self.check_remote_datasets(requests).await;
+
+        if !missed_remote_dependencies.is_empty() {
+            let missed_dependiency_warnings: Vec<String> = missed_remote_dependencies
+                .iter()
+                .map(|remote_ref_dependency| {
+                    format!(
+                        "Dataset:\n {}\nDependencies:\n {}",
+                        remote_ref_dependency.source_dataset,
+                        remote_ref_dependency.dependencies.join("\n"),
+                    )
+                })
+                .collect();
+
+            eprintln!(
+                "{}",
+                console::style(format!(
+                    "Unable verify derivative transformation for some dataset(s). Please download \
+                     next dependencies: \n{}\nor add --integrity flag and try again",
+                    missed_dependiency_warnings.join("\n"),
+                ))
+                .yellow()
+            );
+        }
+
+        if filtered_requests.is_empty() {
             return Ok(vec![]);
         }
 
@@ -99,8 +154,72 @@ impl VerifyCommand {
         Ok(self
             .verification_svc
             .clone()
-            .verify_multi(requests, options, Some(listener))
+            .verify_multi(filtered_requests, options, Some(listener))
             .await)
+    }
+
+    // Return tuple with filtered VerificationRequests(check existing)
+    //   with list of missed remote dependencies
+    async fn check_remote_datasets(
+        &self,
+        verifiation_requests: Vec<VerificationRequest>,
+    ) -> (Vec<VerificationRequest>, Vec<RemoteRefDependency>) {
+        let mut result = vec![];
+        let mut missed_dependencies = vec![];
+
+        for verification_request in &verifiation_requests {
+            if let Ok(dataset_handle) = self
+                .dataset_repo
+                .resolve_dataset_ref(&verification_request.dataset_ref)
+                .await
+            {
+                let is_remote = self
+                    .remote_alias_reg
+                    .get_remote_aliases(&dataset_handle.as_local_ref())
+                    .await
+                    .unwrap()
+                    .get_by_kind(RemoteAliasKind::Pull)
+                    .next()
+                    .is_some();
+                if !is_remote || self.integrity {
+                    result.push(verification_request.clone());
+                    continue;
+                }
+
+                let dataset = self
+                    .dataset_repo
+                    .get_dataset(&dataset_handle.as_local_ref())
+                    .await
+                    .unwrap();
+                let summary = dataset
+                    .get_summary(GetSummaryOpts::default())
+                    .await
+                    .unwrap();
+
+                let mut current_missed_dependencies = vec![];
+
+                for dependecy in summary.dependencies {
+                    if self
+                        .dataset_repo
+                        .resolve_dataset_ref(&DatasetRef::ID(dependecy.clone()))
+                        .await
+                        .is_err()
+                    {
+                        current_missed_dependencies.push(dependecy.to_string());
+                    }
+                }
+                if !current_missed_dependencies.is_empty() {
+                    missed_dependencies.push(RemoteRefDependency {
+                        source_dataset: dataset_handle.alias.to_string(),
+                        dependencies: current_missed_dependencies,
+                    });
+                } else {
+                    result.push(verification_request.clone());
+                }
+            }
+        }
+
+        (result, missed_dependencies)
     }
 }
 
@@ -110,12 +229,6 @@ impl Command for VerifyCommand {
         if self.refs.len() != 1 {
             return Err(CLIError::usage_error(
                 "Verifying multiple datasets at once is not yet supported",
-            ));
-        }
-
-        if self.recursive {
-            return Err(CLIError::usage_error(
-                "Verifying datasets recursively is not yet supported",
             ));
         }